plotting changes
LFP and aMUA plotting code are different in how they attach the y-axis labeling. LFP uses yticks, and this works because the LFP range can accomodate a large tick range. aMUA could not accomodate a large tick range, so I got rid of the ticks and created labels. This is better overall, but still produces a noticeable different at the 12th contact plot.